src/server/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiokafka==0.12.0
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
asyncio==3.4.3
attrs==25.3.0
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
docker==7.1.0
fastapi==0.115.12
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
multidict==6.2.0
packaging==25.0
pluggy==1.5.0
propcache==0.3.1
psycopg2-binary==2.9.10
pyaes==1.6.1
pyasn1==0.6.1
pybreaker==1.3.0
pydantic==2.11.2
pydantic-settings==2.8.1
pydantic_core==2.33.1
pytest==8.3.5
pytest-asyncio==0.26.0
python-dotenv==1.1.0
pytz==2025.2
redis==6.0.0
requests==2.32.3
rsa==4.9
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.40
starlette==0.46.1
Telethon==1.39.0
testcontainers==4.10.0
typing-inspection==0.4.0
typing_extensions==4.13.0
urllib3==2.3.0
uvicorn==0.34.0
wrapt==1.17.2
yarl==1.18.3
